[NOW-149] Add wired and wireless pairing from topology view
- Correct the typo wired
- Move StatefulPopupMenuItem to a file
- Change the function name _handleNodeAction to _handleSelectedNodeAction
